pyproject.toml
[project]
name = "cpio-engagement-engine"
version = "0.1.0"
description = "CryptoPrism Social Engagement Engine — Discord bot loop"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "python-dotenv>=1.0",
    "apify-client>=1.6",
    "aiohttp>=3.9",
]
